+++ webapi/src/test/scala/org/knora/webapi/messages/util/rdf/KnoraResponseV2Spec.scala
@@ -131,7 +131,7 @@ abstract class KnoraResponseV2Spec(featureToggle: FeatureToggle) extends CoreSpe
   "KnoraResponseV2" should {
     "convert Turtle to JSON-LD" in {
       // Read a Turtle file representing a resource. TODO: Use sample project metadata for this test.
-      val turtle: String = FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl"))
+      val turtle: String = FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl"))
 
       // Wrap it in a KnoraTurtleResponseV2.
       val turtleTestResponse = TurtleTestResponse(turtle)
@@ -150,7 +150,7 @@ abstract class KnoraResponseV2Spec(featureToggle: FeatureToggle) extends CoreSpe
 
       // Read an isomorphic JSON-LD file and parse it to a JsonLDDocument.
       val expectedJsonLD: String =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
       val parsedExpectedJsonLD: JsonLDDocument = JsonLDUtil.parseJsonLD(expectedJsonLD)
 
       // Compare the two documents.
@@ -159,7 +159,7 @@ abstract class KnoraResponseV2Spec(featureToggle: FeatureToggle) extends CoreSpe
 
     "convert JSON-LD to Turtle" in {
       // Read a JSON-LD file representing a resource.
-      val jsonLD: String = FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
+      val jsonLD: String = FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
 
       // Wrap it in a KnoraJsonLDResponseV2.
       val jsonLDTestResponse = JsonLDTestResponse(JsonLDUtil.parseJsonLD(jsonLD))
@@ -178,7 +178,7 @@ abstract class KnoraResponseV2Spec(featureToggle: FeatureToggle) extends CoreSpe
 
       // Read an isomorphic Turtle file and parse it to an RDF4J Model.
       val expectedTurtle: String =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl"))
       val parsedExpectedTurtle: RdfModel = rdfFormatUtil.parseToRdfModel(rdfStr = expectedTurtle, rdfFormat = Turtle)
 
       // Compare the two models.